[TC] fix exception

+++ thankscard-adr/src/main/java/trente/asia/thankscard/services/mypage/MypageFragment.java
@@ -81,8 +81,8 @@ public class MypageFragment extends AbstractTCFragment{
 		float screenWidth = Float.parseFloat(preference.get(TcConst.PREF_FRAME_WIDTH));
 		float normalMessageWidth = screenWidth - WelfareUtil.dpToPx(140);
 		float photoMessageWidth = screenWidth / 2 - WelfareUtil.dpToPx(20);
-		float normalTextSize = normalMessageWidth / 15;
-		float photoTextSize = photoMessageWidth / 15;
+		int normalTextSize = (int)(normalMessageWidth / 15);
+		int photoTextSize = (int)(photoMessageWidth / 15);
 		preference.set(TcConst.PREF_NORMAL_TEXT_SIZE, String.valueOf(normalTextSize));
 		preference.set(TcConst.PREF_PHOTO_TEXT_SIZE, String.valueOf(photoTextSize));
 		checkBirthday(preference);
